+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.2.0] - December 29, 2020 (Kent, CT)
+- decitala.utils now includes ``frame_is_spanned_by_slur`` function.
 - Database now notes music21 spanners in frames of ``decitala.trees.rolling_search``.
 
 ## [v0.1.1] - December 28, 2020 (Kent, CT)